Organize Class Files
Say goodbye to Self.cs since its now been moved and renamed.

Moved Other Functions into Proper Locations
