+++ GameLauncher/App/Classes/LauncherCore/APICheckers/APIStatusChecker.cs
@@ -1,3 +1,4 @@
+using GameLauncher.App.Classes.LauncherCore.Global;
 using GameLauncher.App.Classes.Logger;
 using System;
 using System.Net;
@@ -7,7 +8,7 @@ namespace GameLauncher.App.Classes
 {
     class APIStatusChecker
     {
-        public static API CheckStatus(string APIURI)
+        public static APIStatus CheckStatus(string APIURI)
         {
             if (!string.IsNullOrEmpty(APIURI))
             {
@@ -21,7 +22,7 @@ namespace GameLauncher.App.Classes
                     requestAPIStatus.UserAgent = "GameLauncher " + Application.ProductVersion + " (+https://github.com/SoapBoxRaceWorld/GameLauncher_NFSW)";
                     serverResponse = (HttpWebResponse)requestAPIStatus.GetResponse();
                     Log.Info("CORE: " + APIURI + " is Online!");
-                    return API.Online;
+                    return APIStatus.Online;
                     //Do something with response.Headers to find out information about the request
                 }
                 catch (WebException e)
@@ -35,51 +36,51 @@ namespace GameLauncher.App.Classes
 
                         if ((int)serverResponse.StatusCode == 404)
                         {
-                            return API.NotFound;
+                            return APIStatus.NotFound;
                         }
                         else if ((int)serverResponse.StatusCode == 500)
                         {
-                            return API.ServerError;
+                            return APIStatus.ServerError;
                         }
                         else if ((int)serverResponse.StatusCode == 502)
                         {
-                            return API.ServerOverloaded;
+                            return APIStatus.ServerOverloaded;
                         }
                         else if ((int)serverResponse.StatusCode == 503)
                         {
-                            return API.ServerUnavailable;
+                            return APIStatus.ServerUnavailable;
                         }
                         else if ((int)serverResponse.StatusCode == 504)
                         {
-                            return API.GetWayTimeOut;
+                            return APIStatus.GetWayTimeOut;
                         }
                         else if ((int)serverResponse.StatusCode == 520)
                         {
-                            return API.Unknown;
+                            return APIStatus.Unknown;
                         }
                         else if ((int)serverResponse.StatusCode == 521)
                         {
-                            return API.Offline;
+                            return APIStatus.Offline;
                         }
                         else if ((int)serverResponse.StatusCode == 522)
                         {
-                            return API.ConnectionTimeOut;
+                            return APIStatus.ConnectionTimeOut;
                         }
                         else if ((int)serverResponse.StatusCode == 523)
                         {
-                            return API.OriginUnreachable;
+                            return APIStatus.OriginUnreachable;
                         }
                         else if ((int)serverResponse.StatusCode == 524)
                         {
-                            return API.Timeout;
+                            return APIStatus.Timeout;
                         }
                         else if ((int)serverResponse.StatusCode == 525)
                         {
-                            return API.SSLFailed;
+                            return APIStatus.SSLFailed;
                         }
                         else if ((int)serverResponse.StatusCode == 526)
                         {
-                            return API.InvaildSSL;
+                            return APIStatus.InvaildSSL;
                         }
                         //Set flag if there was a timeout or some other issues
                     }
@@ -89,7 +90,7 @@ namespace GameLauncher.App.Classes
                         Console.Write("Error: {0}", e.Status);
                         Console.Write("------------\n");
                         Log.Error("CORE: " + APIURI + " is Offline!");
-                        return API.Offline;
+                        return APIStatus.Offline;
                     }
                 }
                 finally
@@ -100,27 +101,7 @@ namespace GameLauncher.App.Classes
                     }
                 }
             }
-            return API.Null;
+            return APIStatus.Null;
         }
-    }
-    enum API
-    {
-        Offline,
-        Online,
-        BadRequest,
-        Forbidden,
-        NotFound,
-        NotImplmented,
-        ServerError,
-        ServerOverloaded,
-        ServerUnavailable,
-        GetWayTimeOut,
-        ConnectionTimeOut,
-        OriginUnreachable,
-        Timeout,
-        SSLFailed,
-        InvaildSSL,
-        Unknown,
-        Null
     }
 }
