+++ GameLauncher/App/Classes/LauncherCore/APICheckers/VisualsAPIChecker.cs
@@ -1,3 +1,4 @@
+using GameLauncher.App.Classes.LauncherCore.Global;
 using GameLauncherReborn;
 
 namespace GameLauncher.App.Classes.LauncherCore.APICheckers
@@ -14,9 +15,9 @@ namespace GameLauncher.App.Classes.LauncherCore.APICheckers
 
         public static void PingAPIStatus()
         {
-            switch (APIStatusChecker.CheckStatus(Self.mainserver + "/serverlist.json"))
+            switch (APIStatusChecker.CheckStatus(URLs.mainserver + "/serverlist.json"))
             {
-                case API.Online:
+                case APIStatus.Online:
                     break;
                 default:
                     UnitedAPI = false;
@@ -25,9 +26,9 @@ namespace GameLauncher.App.Classes.LauncherCore.APICheckers
 
             if (UnitedAPI == false)
             {
-                switch (APIStatusChecker.CheckStatus(Self.staticapiserver + "/serverlist.json"))
+                switch (APIStatusChecker.CheckStatus(URLs.staticapiserver + "/serverlist.json"))
                 {
-                    case API.Online:
+                    case APIStatus.Online:
                         break;
                     default:
                         CarbonAPI = false;
@@ -37,9 +38,9 @@ namespace GameLauncher.App.Classes.LauncherCore.APICheckers
 
             if (CarbonAPI == false)
             {
-                switch (APIStatusChecker.CheckStatus(Self.secondstaticapiserver + "/serverlist.json"))
+                switch (APIStatusChecker.CheckStatus(URLs.secondstaticapiserver + "/serverlist.json"))
                 {
-                    case API.Online:
+                    case APIStatus.Online:
                         break;
                     default:
                         CarbonAPITwo = false;
@@ -49,15 +50,17 @@ namespace GameLauncher.App.Classes.LauncherCore.APICheckers
 
             if (CarbonAPITwo == false)
             {
-                switch (APIStatusChecker.CheckStatus(Self.woplserver + "/serverlist.json"))
+                switch (APIStatusChecker.CheckStatus(URLs.woplserver + "/serverlist.json"))
                 {
-                    case API.Online:
+                    case APIStatus.Online:
                         break;
                     default:
                         WOPLAPI = false;
                         break;
                 }
             }
+
+            FunctionStatus.IsVisualAPIsChecked = true;
         }
     }
 }
