+++ GameLauncher/App/Classes/LauncherCore/Global/EnumTypes.cs
@@ -1,0 +1,39 @@
+﻿namespace GameLauncher.App.Classes.LauncherCore.Global
+{
+    /* Used with CheckFolder Function in FunctionStatus.cs
+    /* Or
+    * Other Conditionals */
+    enum FolderType
+    {
+        IsTempFolder,
+        IsUsersFolders,
+        IsProgramFilesFolder,
+        IsWindowsFolder,
+        IsSameAsLauncherFolder,
+        Unknown
+    }
+
+    /* Used with CheckStatus Function in APIStatusChecker.cs
+    /* Or
+     * Other Conditionals */
+    enum APIStatus
+    {
+        Offline,
+        Online,
+        BadRequest,
+        Forbidden,
+        NotFound,
+        NotImplmented,
+        ServerError,
+        ServerOverloaded,
+        ServerUnavailable,
+        GetWayTimeOut,
+        ConnectionTimeOut,
+        OriginUnreachable,
+        Timeout,
+        SSLFailed,
+        InvaildSSL,
+        Unknown,
+        Null
+    }
+}
