+++ GameLauncher/App/Classes/LauncherCore/Global/FunctionStatus.cs
@@ -1,0 +1,84 @@
+﻿using System;
+using System.IO;
+using System.Net;
+using System.Windows.Forms;
+
+namespace GameLauncher.App.Classes.LauncherCore.Global
+{
+    class FunctionStatus
+    {
+        /* ServerList Load Checks */
+        public static string ServerListStatus = "Unknown";
+
+        /* CDNList Load Checks */
+        public static string CDNListStatus = "Unknown";
+
+        /* System Language */
+        public static string CurrentLanguage = "EN";
+
+        /* User's Discord ID */
+        public static string DiscordUserID = String.Empty;
+
+        /* Sets Game Launchers User Agent (If Required) */
+        public static string UserAgent = null;
+
+
+        /* Verify Hash Status */
+        public static bool IsVerifyHashDisabled = false;
+
+        /* Visual API Status */
+        public static bool IsVisualAPIsChecked = false;
+
+        /* Sets Conditional to If its Possible to Close Game */
+        public static Boolean CanCloseGame = true;
+
+
+        /* Checks if we have Write Permissions */
+        public static bool HasWriteAccessToFolder(string path)
+        {
+            try
+            {
+                File.Create(path + "temp.txt").Close();
+                File.Delete(path + "temp.txt");
+            }
+            catch
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        /* Used to Center WinForms Forms (Parent Screen) */
+        public static void CenterScreen(Form form)
+        {
+            form.StartPosition = FormStartPosition.Manual;
+            form.Top = (Screen.PrimaryScreen.Bounds.Height - form.Height) / 2;
+            form.Left = (Screen.PrimaryScreen.Bounds.Width - form.Width) / 2;
+        }
+
+        /* Check if Folder Location is Acceptable and Returns a Value
+        /* Let's actually make it cleaner and nicer - MeTonaTOR */
+        public static FolderType CheckFolder(string FolderName)
+        {
+            if (FolderName.Contains("C:\\Users") && FolderName.Contains("Temp")) return FolderType.IsTempFolder;
+            if (FolderName.Contains("C:\\Users")) return FolderType.IsUsersFolders;
+            if (FolderName.Contains("C:\\Program Files")) return FolderType.IsProgramFilesFolder;
+            if (FolderName.Contains("C:\\Windows")) return FolderType.IsWindowsFolder;
+            if (FolderName + "\\" == AppDomain.CurrentDomain.BaseDirectory) return FolderType.IsSameAsLauncherFolder;
+
+            return FolderType.Unknown;
+        }
+
+        /* Converts Host Name to a IP (ex. http://localhost -> 192.168.1.69 */
+        public static string HostName2IP(string hostname)
+        {
+            IPHostEntry iphost = Dns.GetHostEntry(hostname);
+            IPAddress[] addresses = iphost.AddressList;
+            return addresses[0].ToString();
+        }
+
+        /* Moved "runAsAdmin" Code to Gist */
+        /* https://gist.githubusercontent.com/DavidCarbon/97494268b0175a81a5f89a5e5aebce38/raw/eec2f9f80aa4b350ab98d32383e1ee1f2e1c26fd/Self.cs */
+    }
+}
