+++ GameLauncher/App/Classes/LauncherCore/Lists/LanguageListUpdater.cs
@@ -1,10 +1,9 @@
 using GameLauncher.App.Classes.Logger;
-using GameLauncherReborn;
 using Newtonsoft.Json;
 using System;
-using System.Net;
 using System.Collections.Generic;
 using System.Linq;
+using GameLauncher.App.Classes.LauncherCore.Lists.JSON;
 
 namespace GameLauncher.App.Classes
 {
