+++ GameLauncher/GameLauncher.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -160,21 +160,27 @@
     <Compile Include="App\Classes\LauncherCore\Client\Auth\Tokens.cs" />
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\FileAccountSave.cs" />
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\FileSettingsSave.cs" />
+    <Compile Include="App\Classes\LauncherCore\Global\EnumTypes.cs" />
+    <Compile Include="App\Classes\LauncherCore\Global\FunctionStatus.cs" />
+    <Compile Include="App\Classes\LauncherCore\Global\URLs.cs" />
     <Compile Include="App\Classes\LauncherCore\Hash\SHATwoFiveSix.cs" />
-    <Compile Include="App\Classes\LauncherCore\Lists\JSON\CDN.cs" />
-    <Compile Include="App\Classes\LauncherCore\Lists\JSON\LanguageScheme.cs" />
+    <Compile Include="App\Classes\LauncherCore\Lists\JSON\JsonCDN.cs" />
+    <Compile Include="App\Classes\LauncherCore\Lists\JSON\JsonLanguage.cs" />
+    <Compile Include="App\Classes\LauncherCore\Lists\JSON\JsonRootCA.cs" />
+    <Compile Include="App\Classes\LauncherCore\Lists\JSON\JsonServerInfo.cs" />
     <Compile Include="App\Classes\LauncherCore\Lists\LanguageListUpdater.cs" />
     <Compile Include="App\Classes\LauncherCore\ModNet\ModNetLinksCleanup.cs" />
-    <Compile Include="App\Classes\LauncherCore\Validator\CodeSign\JSONRootCA.cs" />
     <Compile Include="App\Classes\LauncherCore\Validator\Email\IsEmailValid.cs" />
     <Compile Include="App\Classes\LauncherCore\Validator\JSON\IsJSONValid.cs" />
-    <Compile Include="App\Classes\LauncherCore\Visuals\Delay.cs" />
     <Compile Include="App\Classes\LauncherCore\Visuals\ProgressBar.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="App\Classes\LauncherCore\Visuals\Theming.cs" />
+    <Compile Include="App\Classes\LauncherCore\Visuals\Time.cs" />
     <Compile Include="App\Classes\LauncherCore\Visuals\TimeConversions.cs" />
     <Compile Include="App\Classes\LauncherCore\Visuals\VerticalBanners.cs" />
+    <Compile Include="App\Classes\SystemPlatform\Components\HardwareID.cs" />
+    <Compile Include="App\Classes\SystemPlatform\Components\HardwareInfo.cs" />
     <Compile Include="App\Classes\SystemPlatform\Linux\DetectLinux.cs" />
     <Compile Include="App\Classes\LauncherCore\Downloader\Downloader.cs" />
     <Compile Include="App\Classes\LauncherCore\Downloader\DownloaderCommand.cs" />
@@ -190,10 +196,8 @@
     <Compile Include="App\Classes\LauncherCore\Downloader\VerifyCommand.cs" />
     <Compile Include="App\Classes\LauncherCore\LauncherUpdater\UpdaterSchema.cs" />
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\ExtractResource.cs" />
-    <Compile Include="App\Classes\SystemPlatform\GPU\init.cs" />
     <Compile Include="App\Classes\LauncherCore\Hash\MD5.cs" />
     <Compile Include="App\Classes\LauncherCore\Hash\SHA.cs" />
-    <Compile Include="App\Classes\SystemPlatform\HWID.cs" />
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\IniFile.cs" />
     <Compile Include="App\Classes\LauncherCore\InsiderKit\EnableInsider.cs" />
     <Compile Include="App\Classes\LauncherCore\Logger\LogToFile.cs" />
@@ -211,13 +215,13 @@
     <Compile Include="App\Classes\LauncherCore\Proxy\ProxyException.cs" />
     <Compile Include="App\Classes\LauncherCore\Proxy\ProxyHandler.cs" />
     <Compile Include="App\Classes\SystemPlatform\Windows\FileORFolderPermissions.cs" />
+    <Compile Include="App\Classes\SystemPlatform\Windows\ManagementSearcher.cs" />
     <Compile Include="App\Classes\SystemPlatform\Windows\RedistributablePackage.cs" />
     <Compile Include="App\Classes\RemovedClasses.cs" />
     <Compile Include="App\Classes\LauncherCore\RPC\CarList.cs" />
     <Compile Include="App\Classes\LauncherCore\RPC\DiscordGamePresence.cs" />
     <Compile Include="App\Classes\LauncherCore\RPC\EventList.cs" />
     <Compile Include="App\Classes\SystemPlatform\Windows\CertificateStore.cs" />
-    <Compile Include="App\Classes\SystemPlatform\Windows\WindowsDefender.cs" />
     <Compile Include="App\SettingsScreen.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -230,7 +234,6 @@
     <Compile Include="App\SplashScreen.Designer.cs">
       <DependentUpon>SplashScreen.cs</DependentUpon>
     </Compile>
-    <Compile Include="App\Classes\LauncherCore\Self.cs" />
     <Compile Include="App\Classes\LauncherCore\Lists\CDNListUpdater.cs" />
     <Compile Include="App\Classes\LauncherCore\Lists\ServerListUpdater.cs" />
     <Compile Include="App\Classes\LauncherCore\Proxy\ServerProxy.cs" />
@@ -254,7 +257,6 @@
     <Compile Include="App\MainScreen.Designer.cs">
       <DependentUpon>MainScreen.cs</DependentUpon>
     </Compile>
-    <Compile Include="App\Classes\LauncherCore\Lists\JSON\JsonScheme.cs" />
     <Compile Include="App\About.cs">
       <SubType>Form</SubType>
     </Compile>
